+++ app/Domain/UseCases/User/UpdateUser.php
@@ -2,37 +2,46 @@
 
 namespace App\Domain\UseCases\User;
 
-use App\Core\Domain\IUseCase;
 use App\Core\Logic\Result;
 use App\Domain\UseCases\Base\UpdateRecord;
+use App\Domain\UseCases\UseCase;
 use App\Presentation\Interfaces\IUserRepository;
 use Exception;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 
-class UpdateUser implements IUseCase
+class UpdateUser extends UseCase
 {
     public function __construct(
         readonly IUserRepository $repository
     )
     {
+        parent::__construct($repository);
     }
 
-    public function execute($data): Result
+    public function execute(): Result
     {
         try {
-            $user = $this->repository->getById($data['id']);
+            $id = $this->getArg('id');
 
-            if($user->isNothing()) {
+            if ($id === null) {
+                throw new Exception('Id is required');
+            }
+
+            $user = $this->repository->getById($id);
+
+            if ($user->isNothing()) {
                 throw new Exception('User not found');
             }
 
+            $data = $this->getArgs();
+
             // check for photo deletion
-            if(isset($data['remove_photo']) && $data['remove_photo'] || $data['photo']) {
+            if (isset($data['remove_photo']) && $data['remove_photo'] || $data['photo']) {
                 // delete old photo
                 if ($photo = $user->get()->photo) {
                     $path = str_replace('storage', 'public', $photo);
-                    if(Storage::exists($path)) {
+                    if (Storage::exists($path)) {
                         Storage::delete($path);
                     }
                 }
@@ -44,7 +53,9 @@ class UpdateUser implements IUseCase
                 $data['photo'] = str_replace('public', 'storage', $data['photo']->storeAs('public/profile/images', $filename));
             }
 
-            $result = UpdateRecord::create($this->repository)->execute($data);
+            $result = UpdateRecord::create($this->repository)
+                ->setArgs($data)
+                ->execute();
         } catch (Exception $e) {
             $result = Result::from($e);
         }
